Moved ArrayList to its own module.
